# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "A BCF API"
  version: "1.0.0"
  title: "BCFAPI"
host: "virtserver.swaggerhub.com"
basePath: "/swisx/bcfAPI/1.0.0"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /bcf/{version}/projects/{project_id}/topics/events:
    get:
      tags:
      - "topic events"
      summary: "Retrieve a collection of topic events related to a project."
      description: "Retrieve a collection of topic events related to a project (default\
        \ sort order is date)."
      operationId: "bcfVersionProjectsProject_idTopicsEventsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/topic_event_GET"
          examples:
            application/json:
            - topic_guid: "A211FCC2-3A3B-EAA4-C321-DE22ABC8414"
              date: "2014-11-19T14:24:11.316Z"
              author: "Architect@example.com"
              events:
              - type: "status_updated"
                value: "Closed"
            - topic_guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              date: "2013-10-21T17:34:22.409Z"
              author: "Architect@example.com"
              events:
              - type: "type_updated"
                value: "Warning"
      x-swagger-router-controller: "TopicEvents"
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/events:
    get:
      tags:
      - "topic events"
      summary: "Retrieve a collection of topic events related to a project."
      description: "Retrieve a collection of topic events related to a project (default\
        \ sort order is date)."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidEventsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/topic_event_GET"
          examples:
            application/json:
            - topic_guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              date: "2014-11-19T14:24:11.316Z"
              author: "Architect@example.com"
              events:
              - type: "type_updated"
                value: "Error"
            - topic_guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              date: "2013-10-21T17:34:22.409Z"
              author: "Architect@example.com"
              events:
              - type: "status_updated"
                value: "Open"
      x-swagger-router-controller: "TopicEvents"
...
definitions:
...
  topic_event_GET:
    type: "object"
    required:
    - "author"
    - "date"
    - "topic_guid"
    properties:
      topic_guid:
        type: "string"
      date:
        type: "string"
      author:
        type: "string"
      actions:
        type: "array"
        items:
          $ref: "#/definitions/event_action"
        minItems: 1
...
  event_action:
    type: "object"
    required:
    - "type"
    properties:
      type:
        type: "string"
      value:
        type: "string"
...